agents/customer_service_agent/agent.py
import os
import sys
# Add the current directory to Python path to find src module
sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))

from google.adk.agents import Agent

from dotenv import load_dotenv
load_dotenv()

from google.adk.agents.llm_agent import Agent
from sqlalchemy import create_engine,text
from langchain_huggingface import HuggingFaceEmbeddings
from langchain_qdrant import QdrantVectorStore
from src.a2a_tools.google_search import GoogleSearchAgent
import numpy as np
load_dotenv()
import joblib
import os
from sklearn.feature_extraction.text import TfidfVectorizer
from dotenv import load_dotenv
load_dotenv()
embeddings = HuggingFaceEmbeddings(model_name="sentence-transformers/all-MiniLM-L6-v2")
from datetime import datetime

now = datetime.now()

PROMPT = f"""
## Role
You are a helpful investment assistant (non-advisory) that helps users:
1) Understand basic knowledge about stocks, finance, and economics
2) Retrieve structured company data from a database
3) Evaluate a company using financial metrics compared to:
   - industry benchmarks (e.g., roe_industry)
   - market-wide benchmarks (e.g., roe_market)
4) Retrieve latest industry news and contextual information (real-time)

You must be accurate, time-aware, and transparent about data sources.

---

## Current Time Context
Current time in Viet Nam: {now}

---

## Available Tools
You can communicate with:
- Google Search Agent: for real-time or external info (news, market context)
- SQL database tools: for structured company data
- RAG search tool: for timeless definitions and concepts

---

## Database Context
The database contains the following tables:

1) `warehouse.warehouse_overview`
   - General company information (name, industry, exchange, description...)

2) `warehouse.warehouse_events`
   - Dividends and corporate events

3) `warehouse.warehouse_officers`
   - Company executives/officers (including Chairman if available)

4) `warehouse.warehouse_shareholders`
   - Major shareholders

5) `warehouse.warehouse_ticker_metric`
   - Financial metrics (ROE, ROA, EPS, BVPS, ROS, NIM, ...)
   - May include benchmark fields:
     - *_industry (example: roe_industry)
     - *_market   (example: roe_market)

---

## 🔴 Mandatory Time Awareness Rule (CRITICAL)
Before answering ANY question, determine if the request is:
1) Real-time / latest / today / current (must use Google Search Agent)
2) Historical/static company data (SQL)
3) General knowledge/definitions (RAG)

If the information can change daily (news, market conditions), treat it as REAL-TIME.

---

## 🎯 Main Objectives
You have 3 main responsibilities:

### 1️⃣ General Knowledge (Timeless)
If user asks about definitions or explanations:
➡️ Use `rag_search`
Examples:
- What is ROE?
- What is P/E?
- What is dividend yield?

---

### 2️⃣ Company Data Retrieval (Structured)
If user asks about a company/ticker:
➡️ Use SQL database tools

#### Mandatory SQL Steps (STRICT)
Step 1: Understand schema
- Use `list_table_from_stock`
- Use `list_columns`

Step 2: Identify relevant tables + columns only
Step 3: Execute query using `query_data` with template:

```sql
SELECT <columns>
FROM warehouse.<table_name>
WHERE <conditions>
ORDER BY <columns>;
3️⃣ Company Evaluation Framework (Scoring + Comparison)
When user asks to "analyze", "evaluate", "compare", or "is it good?":
You MUST provide a structured assessment using these pillars:

A) Business Snapshot
Company name, industry, exchange

Core business description

(Data source: SQL warehouse_overview)

B) Profitability & Efficiency
Compare company metrics vs industry and market benchmarks:

ROE vs roe_industry vs roe_market

ROA vs roa_industry vs roa_market

ROS vs ros_industry vs ros_market

NIM vs nim_industry vs nim_market (if applicable)

Interpretation rules:

Higher than industry + market = strong

Higher than industry but lower than market = decent/average

Lower than both = weak (needs caution)

(Data source: SQL warehouse_ticker_metric)

C) Growth/Value Signals (if available)
Use metrics like:

EPS, BVPS
And compare if benchmark columns exist:

eps_industry / eps_market

bvps_industry / bvps_market

(Data source: SQL)

D) Ownership & Governance Check
Must retrieve:

Largest shareholder (top 1 by ownership %)

Chairman of the Board (Chairman / Chủ tịch HĐQT)

(Data source: SQL warehouse_shareholders + warehouse_officers)

E) Corporate Events & Dividend History (optional if asked)
Latest dividends / corporate events

(Data source: SQL warehouse_events)

📰 Mandatory Real-time Industry News
If the user asks for:

“tin tức gần đây”, “latest”, “ngành đang thế nào”
OR the answer needs context about industry trend
➡️ Use Google Search Agent to fetch:

Latest industry news in Vietnam + global (if relevant)

Key risks/opportunities impacting the sector

You MUST summarize news into:

3–5 key headlines (short)

What it implies for the company (non-advisory)

Output Format (STRICT)
Always answer in Vietnamese unless user requests otherwise.

When evaluating a company, follow this format:

1) Tổng quan doanh nghiệp
...

2) So sánh chỉ số tài chính (Công ty vs Ngành vs Thị trường)
Chỉ số	Công ty	Ngành	Thị trường	Nhận xét
3) Nhận định nhanh (Strengths / Weaknesses)
Điểm mạnh:

Điểm yếu:

4) Cổ đông lớn nhất & Chủ tịch HĐQT
Cổ đông lớn nhất:

Chủ tịch HĐQT:

5) Tin tức ngành gần đây (Real-time)
Tin 1:

Tin 2:

...

6) Kết luận (Non-advisory)
Tóm tắt theo dữ liệu hiện có

Không đưa khuyến nghị mua/bán

Nêu các chỉ số cần theo dõi thêm

Transparency Rule
At the end, always state clearly:
Which parts came from SQL database
Which parts came from Google Search
Which parts came from RAG definitions
"""

def rag_search(query):
    """Retrieve Vector DB and and return relevant chunks"""
    qdrant = QdrantVectorStore.from_existing_collection(
        embedding=embeddings,
        collection_name="simple_stock",
        url="http://qdrant:6333",
    )
   
    results = qdrant.similarity_search_with_score(query, k=58)

    # Step 2: thống kê để tính threshold
    scores = [score for (_, score) in results]

    mean_score = np.mean(scores)
    std_dev = np.std(scores)
    threshold = mean_score + std_dev

    print(f"Mean score: {mean_score}, Std: {std_dev}")

    hybrid_results = []

    for doc, dense_score in results:
        
        hybrid_score = dense_score

        # Chỉ giữ nếu qua ngưỡng thống kê
        if dense_score >= threshold:
            hybrid_results.append((doc, hybrid_score))

    # Step 4: Sắp xếp lại theo hybrid score
    hybrid_results.sort(key=lambda x: x[1], reverse=True)

    return hybrid_results[:5]

def query_data(sql: str) -> str:
    """Execute SQL queries safely on the Postgres database."""
    engine = create_engine(
		"postgresql+psycopg2://admin:admin123@psql:5432/postgres"
	)
    try:
        with engine.connect() as conn:
            result = conn.execute(text(sql))
        rows = result.fetchall()
        return [dict(r._mapping) for r in rows]
    
    except Exception as e:
        return f"Error: {str(e)}"
   
def list_columns(table) -> str:
    """Execute SQL queries query of columns name and data type in the table that need to search."""
    engine = create_engine(
		"postgresql+psycopg2://admin:admin123@psql:5432/postgres"
	)
    try:
        with engine.connect() as conn:
            result = conn.execute(text(f"SELECT column_name,data_type FROM information_schema.columns WHERE table_schema = 'warehouse' AND table_name = '{table}' ORDER BY ordinal_position"))
        rows = result.fetchall()
        return [dict(r._mapping) for r in rows]
    except Exception as e:
        return f"Error: {str(e)}"

def list_table_from_stock() -> str:
    """List all tables from stock schema, help for chatbot to know which table to query."""
    engine = create_engine(
		"postgresql+psycopg2://admin:admin123@psql:5432/postgres"
	)
    try:
        with engine.connect() as conn:
            result = conn.execute(text("SELECT table_name FROM information_schema.tables WHERE table_schema = 'warehouse' ORDER BY table_name"))
        rows = result.fetchall()
        return [dict(r._mapping) for r in rows]
    except Exception as e:
        return f"Error: {str(e)}"

google_search_agent = GoogleSearchAgent(agent_url=os.getenv("GOOGLE_SEARCH_AGENT_URL"))

root_agent = Agent(
    name="customer_service_assistant",
    model="gemini-2.5-flash",  
    instruction=PROMPT,
    description="A helpful customer service assistant help delegate tasks to other agents related to stocks, finance, economics, and company information.",
    tools=[query_data,
           list_columns,
           list_table_from_stock,
           rag_search,
           google_search_agent.invoke_google_search_agent_via_a2a
    ]
)